Divide bot.rb into two files
